+++ client/src/lib/src/BitbakeSettings.ts
@@ -3,8 +3,6 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import * as vscode from 'vscode'
-
 import path from 'path'
 
 /// Defines the context of a bitbake workspace with all information to call bitbake
@@ -14,25 +12,17 @@ export interface BitbakeSettings {
   pathToEnvScript: string
 }
 
-export function loadBitbakeSettings (): BitbakeSettings {
-  const settings = vscode.workspace.getConfiguration('bitbake')
-
-  const workspaceFolder = vscode.workspace.workspaceFolders?.[0]
-  let workspacePath = ''
-  if (workspaceFolder !== undefined) {
-    workspacePath = workspaceFolder.uri.fsPath
-  }
-
+export function loadBitbakeSettings (settings: any, workspaceFolder: string = ''): BitbakeSettings {
   /* eslint no-template-curly-in-string: "off" */
   // The default values are defined in package.json
-  let pathToBitbakeFolder: string = settings.get<string>('pathToBitbakeFolder', '')
-  pathToBitbakeFolder = pathToBitbakeFolder.replace('${workspaceFolder}', workspacePath)
+  let pathToBitbakeFolder: string = settings.pathToBitbakeFolder
+  pathToBitbakeFolder = pathToBitbakeFolder.replace('${workspaceFolder}', workspaceFolder)
   pathToBitbakeFolder = path.resolve(pathToBitbakeFolder)
-  let pathToBuildFolder: string = settings.get<string>('pathToBuildFolder', '')
-  pathToBuildFolder = pathToBuildFolder.replace('${workspaceFolder}', workspacePath)
+  let pathToBuildFolder: string = settings.pathToBuildFolder
+  pathToBuildFolder = pathToBuildFolder.replace('${workspaceFolder}', workspaceFolder)
   pathToBuildFolder = path.resolve(pathToBuildFolder)
-  let pathToEnvScript: string = settings.get<string>('pathToEnvScript', '')
-  pathToEnvScript = pathToEnvScript.replace('${workspaceFolder}', workspacePath)
+  let pathToEnvScript: string = settings.pathToEnvScript
+  pathToEnvScript = pathToEnvScript.replace('${workspaceFolder}', workspaceFolder)
   pathToEnvScript = path.resolve(pathToEnvScript)
 
   return {
